Rename mod to native
