+++ rust/geodatafusion/src/udf/mod.rs
@@ -1,3 +1,3 @@
-pub mod geo;
 pub mod geos;
 pub mod io;
+pub mod native;
