+++ rust/geodatafusion/src/udf/native/accessors/line_string.rs
