+++ rust/geodatafusion/src/udf/native/bounding_box/box.rs
